updated scroll to top button
