+++ src/routes/+layout.svelte
@@ -4,26 +4,10 @@
 	import Nav from '$lib/components/ui/Nav.svelte';
 	import { Toaster } from '$lib/components/ui/sonner';
 	import SubscribeSection from '$lib/components/ui/SubscribeSection.svelte';
-	import { cn } from '$lib/utils/ui-components';
 	import { injectAnalytics } from '@vercel/analytics/sveltekit';
-	import { ArrowUp } from 'lucide-svelte';
 	import '../app.css';
 
 	injectAnalytics();
-
-	let showButton = $state(false);
-
-	function scrollContainer() {
-		return document.documentElement || document.body;
-	}
-
-	function handleScroll() {
-		showButton = scrollContainer().scrollTop > 600;
-	}
-
-	function handleScrollToTop() {
-		document.body.scrollIntoView({ behavior: 'smooth' });
-	}
 
 	const { children } = $props();
 </script>
@@ -38,8 +22,6 @@
 	<meta name="theme-color" content="#ffffff" />
 </svelte:head>
 
-<svelte:window on:scroll={handleScroll} />
-
 <div>
 	<Nav />
 	<div class="transition-colors duration-300">
@@ -47,27 +29,9 @@
 	</div>
 	<div class="flex flex-col grow items-center">
 		<SubscribeSection />
-		<!-- <Beehiiv /> -->
 		<JoinSection />
 		<Footer />
 		<Toaster richColors position="top-center" />
-
-		<div class="container flex items-center justify-end">
-			<button
-				class={cn(
-					'flex items-center justify-center gap-1 text-2xl group text-[12.667px] size-10 bg-[#19191A] rounded-full text-[#B4B4B4] group fixed bottom-8 md:bottom-11 font-semibold shadow-2xl hover:shadow-hover hover:bg-white border border-[#333] -z-50 transition opacity-0',
-					showButton && 'opacity-100 z-[999999]'
-				)}
-				aria-label="Scroll back to the top of the page"
-				onclick={handleScrollToTop}
-			>
-				<!-- <span class="max-md:hidden">Back to the Top</span> -->
-				<ArrowUp
-					class="size-4 rounded-full group-hover:-translate-y-[2px] will-change-transform transition group-hover:text-black flex-shrink-0"
-					style="stroke-width: 1.4"
-				/>
-			</button>
-		</div>
 	</div>
 </div>
