+++ src/routes/contributors/[slug]/+page.svelte
@@ -5,6 +5,7 @@
 	import Grid from '$lib/components/ui/icons/Grid.svelte';
 	import List from '$lib/components/ui/icons/List.svelte';
 	import X from '$lib/components/ui/icons/X.svelte';
+	import ScrollToTopButton from '$lib/components/ui/ScrollToTopButton.svelte';
 	import type { ArticleMetadata } from '$lib/types/article';
 	import { formatCategorySlug } from '$lib/utils/format';
 	import { cn } from '$lib/utils/ui-components';
@@ -283,3 +284,5 @@
 		{/if}
 	</section>
 </div>
+
+<ScrollToTopButton />
