+++ src/routes/reports/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import BaseHead from '$lib/components/server/BaseHead.svelte';
 	import ArticleList from '$lib/components/ui/ArticleList.svelte';
+	import ScrollToTopButton from '$lib/components/ui/ScrollToTopButton.svelte';
 	import { setArticles } from '$lib/stores/articles.svelte';
 	import { onMount } from 'svelte';
 	import type { PageData } from './$types';
@@ -22,3 +23,5 @@
 	<ArticleList {articles} {articleCategories} articlesPerPage={100} />
 	<!-- <Testimonials5 /> -->
 </div>
+
+<ScrollToTopButton />
